split git commit cache and git commit provider
